+++ src/main/java/br/com/meli/dhprojetointegrador/repository/WarehouseRepository.java
@@ -3,5 +3,5 @@ package br.com.meli.dhprojetointegrador.repository;
 import br.com.meli.dhprojetointegrador.entity.Warehouse;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WarehouseRepository extends JpaRepository<Warehouse, Integer> {
+public interface WarehouseRepository extends JpaRepository<Warehouse, Long> {
 }
