+++ src/main/java/br/com/meli/dhprojetointegrador/service/AgentService.java
@@ -11,7 +11,7 @@ import org.springframework.stereotype.Service;
 public class AgentService {
     private final AgentRepository agentRepository;
 
-    public Agent findAgentById(Integer id) throws BusinessValidatorException {
+    public Agent findAgentById(Long id) throws BusinessValidatorException {
         return agentRepository.findById(id).orElseThrow(() ->
                 new BusinessValidatorException(String.format("Agent with id %d not found", id)));
     }
